[refs #29] Creation of a dedicated Samtools index task
--generic version restricts command line to always having the input file
  as the last argument. this prevents passing in an explicit output name
--samtools can/does detect input type and automatically creates the
  correct index, but the generic task writes this to stdout, which makes
  it hard to pass to downstream tasks

[refs #29] Adds double quotes to the ouput declaration

(p.s. @genomics-geek I've uncovered a womtool/githook loophole
:astonished:)

[refs #29] Forces output file to be set explicitly

while samtools does automatically detect the input filetype, the only
way (discovered after testing) for cromwell to know that this task is
finished is when it detects the output file. if using the default/empty
string, this will never be moved to the execution/ folder